Заставляем классы пули и астероида работать
