+++ AsteroidGame/VisualObjects/Bullet.cs
@@ -12,7 +12,7 @@ namespace AsteroidGame.VisualObjects
         private const int __BulletSizeX = 20;
         private const int __BulletSizeY = 5;
         public Bullet(int Position) 
-            : base(new Point(0,Position), Point.Empty, new Size(__BulletSizeY,__BulletSizeY))
+            : base(new Point(0,Position), Point.Empty, new Size(__BulletSizeX,__BulletSizeY))
         {
         }
 
@@ -21,6 +21,11 @@ namespace AsteroidGame.VisualObjects
             var rect = new Rectangle(_Position,_Size);
             g.FillEllipse(Brushes.Red,rect);
             g.DrawEllipse(Pens.White, rect);
+        }
+
+        public override void Update()
+        {
+            _Position = new Point(_Position.X + 3,_Position.Y);
         }
     }
 }
